Add alternative vectorized behavior of jacobian and hessian across backends + move autodiff tests to test_autodiff
