+++ geomstats/_backend/numpy/autodiff.py
@@ -38,7 +38,27 @@ def jacobian(func):
     )
 
 
+def jacobian_vec(func):
+    """Return an error when using automatic differentiation with numpy."""
+    raise RuntimeError(
+        "Automatic differentiation is not supported with numpy backend. "
+        "Use autograd, pytorch or tensorflow backend instead.\n"
+        "Change backend via the command "
+        "export GEOMSTATS_BACKEND=autograd in a terminal."
+    )
+
+
 def hessian(func):
+    """Return an error when using automatic differentiation with numpy."""
+    raise RuntimeError(
+        "Automatic differentiation is not supported with numpy backend. "
+        "Use autograd, pytorch or tensorflow backend instead.\n"
+        "Change backend via the command "
+        "export GEOMSTATS_BACKEND=autograd in a terminal."
+    )
+
+
+def hessian_vec(func):
     """Return an error when using automatic differentiation with numpy."""
     raise RuntimeError(
         "Automatic differentiation is not supported with numpy backend. "
