+++ geomstats/_backend/tensorflow/autodiff.py
@@ -147,18 +147,58 @@ def jacobian(func):
 
         Here, x is a single point of ndim 1.
 
+        Parameters
+        ----------
+        x : array-like
+            Input to function func or its jacobian.
+
+        Returns
+        -------
+        _ : array-like
+            Value of the jacobian of func at x.
+        """
+        if isinstance(x, _np.ndarray):
+            x = _tf.Variable(x)
+        with _tf.GradientTape() as g:
+            g.watch(x)
+            y = func(x)
+        return g.jacobian(y, x)
+
+    return _jac
+
+
+def jacobian_vec(func):
+    """Return a function that returns the jacobian of func.
+
+    Parameters
+    ----------
+    func : callable
+        Function whose Jacobian is computed.
+
+    Returns
+    -------
+    jac : callable
+        Function taking x as input and returning
+        the jacobian of func at x.
+    """
+
+    def _jac(x):
+        """Return the jacobian of func at x.
+
+        Here, x is a single point of ndim 1.
+
         We note that the jacobian function of torch is not vectorized
         by default, thus we modify its behavior here.
 
         Default tensorflow behavior:
 
-        If the jacobian for one point of shape (2,) is of shape (3, 2),
-        then calling the jacobian on 4 points with shape (4, 2) will
-        be of shape (3, 2, 4, 2).
+        If the jacobian for one point of shape (dim,) is of shape (out_dim, dim),
+        then calling the jacobian on several points with shape (n_points, dim) will
+        be of shape (out_dim, dim, n_points, dim).
 
         Modified behavior:
 
-        Calling the jacobian on 4 points gives a tensor of shape (4, 3, 2).
+        Calling the jacobian on points gives a tensor of shape (n_points, out_dim, dim).
 
         Parameters
         ----------
@@ -200,6 +240,52 @@ def jacobian(func):
 
 
 def hessian(func):
+    """Return a function that returns the hessian of func.
+
+    Parameters
+    ----------
+    func : callable
+        Function whose Hessian is computed.
+
+    Returns
+    -------
+    hess : callable
+        Function taking x as input and returning
+        the hessian of func at x.
+    """
+
+    def _hess(x):
+        """Return the hessian of func at x.
+
+        Parameters
+        ----------
+        x : array-like
+            Input to function func or its hessian.
+
+        Returns
+        -------
+        _ : array-like
+            Value of the hessian of func at x.
+        """
+        # Note: this is a temporary implementation
+        # that uses the jacobian of the gradient.
+        # inspired from https://github.com/tensorflow/tensorflow/issues/29781
+        # waiting for the hessian function to be implemented in GradientTape.
+        if isinstance(x, _np.ndarray):
+            x = _tf.Variable(x)
+
+        with _tf.GradientTape(persistent=True) as g:
+            g.watch(x)
+            y = func(x)
+            grads = g.gradient(y, [x])
+
+        hessians = g.jacobian(grads[0], [x])
+        return hessians[0]
+
+    return _hess
+
+
+def hessian_vec(func):
     """Return a function that returns the hessian of func.
 
     Parameters
